# Documentation/src/content/schemas/swagger.yaml
openapi: 3.0.0
info:
  title: Zello API
  version: 1.0.0
  description: |-
    A comprehensive API for managing projects, lists, and cards similar to Trello. This API allows users to:
    - Create and manage workspaces, projects, lists, and cards
    - Organize work using customizable boards and lists
    - Collaborate with team members through comments and mentions
    - Track work progress with activity logs
    - Manage permissions and roles at different levels
    
    ### Features
    - Full CRUD operations for all resources
    - Rich card features including attachments, checklists, and labels
    - Comprehensive activity tracking
    - Fine-grained permission system
    - Authentication and user management
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    List:
      description: Represents a column in a project (e.g., 'To Do', 'In Progress', 'Done')
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        projectId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        title:
          type: string
          example: In Progress
        position:
          type: integer
          description: Order of the list in the project (0-based)
          example: 1
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
...
    AuthToken:
      description: Authentication tokens for API access
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
    Error:
      type: object
      properties:
        message:
          type: string
          example: Invalid request parameters
        code:
          type: string
          example: INVALID_PARAMETERS
security:
  - bearerAuth: [ ]
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account in the system
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - name
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
            example:
              username: test-user
              email: jane.smith@example.com
              password: securePassword123
              name: Jane Smith
      responses:
        '201':
          description: User successfully registered
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticates a user and returns access and refresh tokens
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: jane.smith@example.com
              password: securePassword123
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Successfully refreshed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: Logout user
      description: Invalidates the current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/password/change:
    post:
      summary: Change password
      description: Change password, token and password are required for password change.
      security: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
            example:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              newPassword: newSecurePassword123
      responses:
        '200':
          description: Password successfully changed
        '400':
          description: Invalid password format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /lists/{listId}:
    get:
      summary: Get list details
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
              examples:
                success:
                  summary: List details response
                  value: {
                    "id": "123e4567-e89b-12d3-a456-426614174002",
                    "projectId": "123e4567-e89b-12d3-a456-426614174030",
                    "title": "In Progress",
                    "position": 1,
                    "settings": {
                      "wipLimit": 5,
                      "color": "#0099FF",
                      "collapsed": false,
                      "description": "Tasks currently being worked on"
                    },
                    "statistics": {
                      "totalCards": 3,
                      "completedCards": 1,
                      "overdueCards": 1
                    },
                    "createdAt": "2024-01-01T12:00:00Z",
                    "updatedAt": "2024-01-15T16:30:00Z",
                    "cards": [
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174060",
                        "title": "Implement user authentication",
                        "priority": "high",
                        "dueDate": "2024-01-20T17:00:00Z",
                        "position": 0
                      },
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174061",
                        "title": "Design landing page",
                        "priority": "medium",
                        "dueDate": "2024-01-25T17:00:00Z",
                        "position": 1
                      },
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174062",
                        "title": "Setup CI/CD pipeline",
                        "priority": "urgent",
                        "dueDate": "2024-01-18T17:00:00Z",
                        "position": 2
                      }
                    ]
                  }
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "List not found"
                code: "LIST_NOT_FOUND"

    put:
      summary: Update list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                position:
                  type: integer
            example:
              title: "In Review"
              position: 3
      responses:
        '200':
          description: List updated
        '403':
          description: Insufficient permissions
        '404':
          description: List not found

    delete:
      summary: Delete list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: List deleted
        '403':
          description: Insufficient permissions
        '404':
          description: List not found
  '/lists/{listId}/cards':
    post:
      summary: Create a new card
      description: Creates a new card (task, issue, etc.) within a list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - type
              properties:
                title:
                  type: string
                type:
                  type: string
                  enum: [ task, bug, feature, improvement, epic ]
                description:
                  type: string
                priority:
                  type: string
                  enum: [ low, medium, high, urgent ]
                severity:
                  type: string
                  enum: [ trivial, minor, major, critical, blocker ]
                startDate:
                  type: string
                  format: date-time
                dueDate:
                  type: string
                  format: date-time
            examples:
              task:
                summary: Create a new task
                value:
                  title: "Implement user authentication"
                  type: "task"
                  description: "Implement OAuth2 authentication flow"
                  priority: "high"
                  startDate: "2024-01-15T00:00:00Z"
                  dueDate: "2024-01-20T23:59:59Z"
                  taskProperties:
                    estimatedHours: 16
                    dependencies: [ ]
              bug:
                summary: Create a new bug report
                value:
                  title: "Login button not working on Safari"
                  type: "bug"
                  description: "Users cannot click the login button in Safari browser"
                  priority: "high"
                  severity: "major"
                  issueProperties:
                    affectedVersion: "1.2.0"
                    environment: "production"
                    browserInfo: "Safari 15.0"
                    stepsToReproduce: "1. Navigate to login page\n2. Fill in credentials\n3. Click login button"
      responses:
        '201':
          description: Card successfully created
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...